pyproject.toml
[project]
name = "src"
version = "0.1.0"
authors = [
    {name = "Amir Mahdi Jamali",email = "mahdijamali101@gmail.com"}
]
readme = "README.md"
description = "Gym management backend system"
requires-python = ">=3.12,<3.14"
dependencies = ["fastapi (>=0.115.12,<0.116.0)", "uvicorn[standard] (>=0.34.3,<0.35.0)", "sqlalchemy (>=2.0.41,<3.0.0)", "alembic (>=1.16.1,<2.0.0)", "psycopg2-binary (>=2.9.10,<3.0.0)", "python-jose[cryptography] (>=3.5.0,<4.0.0)", "passlib[bcrypt] (>=1.7.4,<2.0.0)", "python-multipart (>=0.0.20,<0.0.21)", "pydantic (>=2.11.5,<3.0.0)", "pydantic-settings (>=2.9.1,<3.0.0)", "python-dotenv (>=1.1.0,<2.0.0)", "redis (>=6.2.0,<7.0.0)", "email-validator (>=2.2.0,<3.0.0)", "pillow (>=11.2.1,<12.0.0)", "pytest (>=8.3.5,<9.0.0)", "pytest-asyncio (>=1.0.0,<2.0.0)", "httpx (>=0.28.1,<0.29.0)", "starlette-csrf (>=3.0.0,<4.0.0)", "pytz (>=2025.2,<2026.0)", "slowapi (>=0.1.9,<0.2.0)", "loguru (>=0.7.3,<0.8.0)", "asyncpg (>=0.30.0,<0.31.0)"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"